change behaviour on wrong user input: return a error code instead of raising an exception
When user submits a request having a spacing smaller than the mode minimum spacing,
the consistent check raises an exception for bad service. This behaviour assumes that
the user knows the modes characteristics.
Instead it makes more sense to return an error code in the case when the user doesn't know
about the library details but only on modes names.
this is usefull for CANDI-optical POC use case

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>
Change-Id: Ic7e02566a0cd5750a07752db7c318c29c62b3e1f