Add a frequency-dependent fiber model declared as a list in the equipment-file.
This fiber model is defined in the equipment-file, and can then be invoked at
the topology-file level by simply specifying the 'type_variety'.
Hence it is no longer required to copy the model definition per fiber instance
in the topology-file.

Once instanciated in the topology-file, the model can be customized as follows:
the legacy 'loss_coef'-scalar is used to scale the instantiated fiber model
for the given 'ref_frequency' such that at the
'ref_frequency', the model values the 'loss_coef'-scalar.

As the 'loss_coef'-scalar is declared in the 'topology'-file similarly to the
legacy way, applying different loss coefficient models (or none), is eased,
and done straightforwardly via the 'type_variety'.

In case a legacy dict-model is declared as well as a list-model,
then the legacy model has priority.

Co-authored-by: EstherLerouzic <esther.lerouzic@orange.com>

Signed-off-by: Florian FRANK <florian1.frank@orange.com>
Change-Id: Id4267c2bc78d823031e7a69b11115eb93df39183